[UI][Tizen] Fix TizenSystemIndicator lanscape mode
This patch adds a temporary fix for landscape mode and an initial part
of the real implementation.

In lanscape mode, the indicator in tizen is a little bit different than
in portrait mode, just the battery bar is showed, without a background.
When you drag this part, the rest of the indicator is showed on screen
(same as portrait mode), but in this case, the image is dragged over the
screen content.